Move all the based segments aspects into ISegmentProvider
